+++ hashing/HashingWithChaining.h
@@ -6,15 +6,15 @@
 
 class HashingWithChaining {
 public:
-    explicit HashingWithChaining(uint32_t m, uint8_t w, const std::vector<uint32_t> &list);
-    bool query(uint32_t x);
+    explicit HashingWithChaining(uint32_t m, uint8_t w, const std::vector<uint64_t> &list);
+    bool query(uint64_t x);
     void print_table() const;
     uint32_t largest_linked_list() const;
 
 private:
-    void insert(uint32_t x);
+    void insert(uint64_t x);
 
-    std::vector<std::vector<uint32_t>> table;
+    std::vector<std::vector<uint64_t>> table;
     MultiplyShiftHash h;
 };
