+++ utils/NumberUtils.h
@@ -9,8 +9,7 @@ class NumberUtils {
 public:
   static uint32_t generate_random_uint(uint32_t m);
   static uint64_t generate_random_uint(uint64_t m);
-  static std::vector<uint32_t> generate_shuffled_list(uint32_t size);
-
+  static std::vector<uint64_t> generate_shuffled_list(uint32_t size);
 };
 
 #endif //NUMBERUTILS_H
